requirements.txt
torch==2.3.0
torchvision==0.18.0
Flask==3.0.3
pillow==10.2.0
numpy==1.19.5
pickle4
scikit-learn==0.23.0
gunicorn